Use contents of title.txt as applications name
Current implementation uses demo application's directory name as
a title in the qtlauncher. This is problematic when we want to run a
qmake on a demo and at the same time have a title that consists of
spaces.

Therefore if a demo provides a file called "title.txt" we use its
content as a title.

Change-Id: Id6e2ad94dca3b021acb9ad3ece952f48053e4b03
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>