Fix PieMenu not showing in enterprise-gallery launcher demo.
The patch provided by Mitch.

The problem is that the root item doesn't get the correct size - or
any size for that matter - so it is never visible.

The fix for this is already in Qt 5.3 (see QTBUG-36938), but we can
still work around it for earlier versions with this fix.

Task-number: QTRD-3040
Change-Id: Iaa014624411d321eec6caf30ea0bd4921ff2ae72
Reviewed-by: Kalle Viironen <kalle.viironen@digia.com>
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>