Add possibility to launch QtWebEngine applications
QtWebEngine needs OpenGL resource sharing set up in QtQuick.
This also exposes QUrl::fromUserInput to be used by the browser.
We have to pass --single-process for now since the QML Loader
element is for some reason not able to load a QML file with
a WebView that spawns a secondary process.

Change-Id: I41c78b65f0ac87d8adabccbaf6f858262de87310
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>