Stop animateToCenter when a finger is on the screen
Currently trying to move the ListView while animateToCenter runs
will create a conflict where Flickable tries to move the view on
some frames and the animation on other frames, causing flickering.

Another issue is that the animation will run to the end if the user
puts his finger on the screen, which is not what one would expect.

Put a MouseArea over the ListView to catch events but making sure
that they still get get propagated to the ListView under.

Change-Id: I7afcc6c9217d1938c9ee0b0cfaee763bce6c9e59
Reviewed-by: Rainer Keller <rainer.keller@digia.com>