Do not set a default font size
It is not a good idea to set a default font size if that
font size is going to be in pixel values.  The default
font size should be a point value so that it can be
adjusted based on the DPI of the screen.  The Qt Quick
Controls scale them selves based on the Implicit Height
of the default font size, and by setting the default
font size to a pixel value, the Implicit Height will
always be the same.

This fixes the issue of Qt Quick Controls being too small
on HiDPI screens.

Change-Id: Iaa47a911e5f4526a25bf9eb6b7ba747ebf746a38
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@theqtcompany.com>