Use a smoother looping animation in loading indicator
Previous had an abrupt restart for the next go around. Instead clear the circle
with the same speed as it is filled. To avoid flashing line cap changes, make
sure there's always a bit of the green line available after the start.

endAngle was confusing naming, because it implied that the angle is given in a
similar way as startAngle, 0 being at the top and 180 at the bottom. This was
not the case, as instead it is the length of the arc in angles. Rename it to
spanAngle.

Task-number: QTBUG-70548
Change-Id: Ia9d916cfa120f81e10c00caf6d10c653a9393efc
Reviewed-by: Sami Nurmenniemi <sami.nurmenniemi@qt.io>
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>
Reviewed-by: Kari Hormi <kari.hormi@qt.io>