Parse individual demo.xml files
Instead of single demos.xml for all the demo descriptions, use
individual demo.xml files that are located in the demo folder.
This makes possible to easily select which demos are shown in
a particular device, when unwanted demos can be simple removed
without editing the common xml file.

Actual folder path is used for the location and exclude list
is no longer used.

Task-number: QTBUG-74930
Change-Id: If6819a80650084feb6f3e68d5a5a7a8468394bfa
Reviewed-by: Kari Oikarinen <kari.oikarinen@qt.io>